fix(runtime): make bigint_ops compile (E0282 in bitwise_not) (bd-9vouw.54)
Commit 2888a9c0a, swept onto main by the auto-commit process before it
was built, does not compile:

    error[E0282]: type annotations needed
       --> crates/franken-engine/src/baseline_interpreter/bigint_ops.rs:163:5
    163 |     (-parse(text) - 1).to_string()

`BigInt - {integer}` has several candidate impls. Subtract an explicit
`BigInt::from(1u8)`.

Evidence: this was the only error in `cargo test -p frankenengine-engine`
(r24b, target_dev) at 2888a9c0a. Builds including this line are running
now; the BigInt behavior tests (bigint_arithmetic_bd_9vouw_54) have not
yet been run against it.

+++ crates/franken-engine/src/baseline_interpreter/bigint_ops.rs
@@ -160,7 +160,7 @@ pub(super) fn negate(text: &str) -> String {
 
 /// `~value` (BigInt::bitwiseNOT): `-value - 1`.
 pub(super) fn bitwise_not(text: &str) -> String {
-    (-parse(text) - 1).to_string()
+    (-parse(text) - BigInt::from(1u8)).to_string()
 }
 
 /// Order of two BigInts.
